Add semver and stage filter to SessionsQueryBuilder
